Move playfields to own class.
