+++ osu.Game/GameModes/Play/Taiko/TaikoHitRenderer.cs
@@ -5,14 +5,11 @@ using System;
 using System.Collections.Generic;
 using osu.Framework.Graphics.Sprites;
 using osu.Framework.Graphics;
-using osu.Framework.Graphics.Containers;
-using osu.Framework.Graphics.Drawables;
 using osu.Framework.Graphics.Transformations;
 using osu.Game.Beatmaps.Objects;
 using osu.Game.Beatmaps.Objects.Osu;
 using osu.Game.Beatmaps.Objects.Taiko;
 using OpenTK;
-using OpenTK.Graphics;
 
 namespace osu.Game.GameModes.Play.Taiko
 {
@@ -84,33 +81,6 @@ namespace osu.Game.GameModes.Play.Taiko
 
                 playfield.Add(s);
             }
-        }
-    }
-
-    public class TaikoPlayfield : Container
-    {
-        public TaikoPlayfield()
-        {
-            SizeMode = InheritMode.X;
-            Size = new Vector2(1, 100);
-            Anchor = Anchor.Centre;
-            Origin = Anchor.Centre;
-        }
-
-        public override void Load()
-        {
-            base.Load();
-
-            Add(new Box() { SizeMode = InheritMode.XY, Alpha = 0.5f });
-
-            Add(new Sprite(Game.Textures.Get(@"menu-osu"))
-            {
-                Origin = Anchor.Centre,
-                Scale = 0.2f,
-                PositionMode = InheritMode.XY,
-                Position = new Vector2(0.1f, 0.5f),
-                Colour = Color4.Gray
-            });
         }
     }
 }
