+++ osu.Game/osu.Game.csproj
@@ -75,8 +75,10 @@
     <Compile Include="GameModes\FieldTest.cs" />
     <Compile Include="GameModes\Play\HitRenderer.cs" />
     <Compile Include="GameModes\Play\Osu\OsuHitRenderer.cs" />
+    <Compile Include="GameModes\Play\Osu\OsuPlayfield.cs" />
     <Compile Include="GameModes\Play\PlayTest.cs" />
     <Compile Include="GameModes\Play\Taiko\TaikoHitRenderer.cs" />
+    <Compile Include="GameModes\Play\Taiko\TaikoPlayfield.cs" />
     <Compile Include="Graphics\Containers\OsuComponent.cs" />
     <Compile Include="Graphics\Containers\OsuGameMode.cs" />
     <Compile Include="Graphics\Containers\OsuLargeComponent.cs" />
